@@ -18,7 +18,7 @@ description: "A new Flutter module project."
 version: 1.0.0+1
 
 environment:
-  sdk: '3.2.4'
+  sdk: '3.3.4'
 
 dependencies:
   flutter:
