@@ -10,7 +10,7 @@ const double left_margin_forget_password_button = 190;
 const String title_appbar = "Đăng ký tài khoản";
 const String hint_text_last_name = "Họ";
 const String hint_text_first_name = "Tên";
-const String hint_text_phone = "Số điện thoại";
+const String hint_text_email = "Email";
 const String hint_text_password = "Mật khẩu";  
 const String text_button = "Đăng ký";
 const String text_button_forget_password = "Quên mật khẩu ?";
